+++ src/main/java/open/commons/spring/web/handler/HttpRequestProxyHeader.java
@@ -26,7 +26,7 @@
 
 package open.commons.spring.web.handler;
 
-import javax.validation.constraints.NotEmpty;
+import java.util.HashMap;
 
 /**
  * Proxy 서버를 통해서 전달되는 실제 클라이언트의 Http 요청 정보.<br>
@@ -34,55 +34,74 @@ import javax.validation.constraints.NotEmpty;
  * 서비스 설정에 아래와 같이 추가합니다.
  * 
  * <pre>
- * spring:
+ * server:
  *   # 헤더 처리는 프레임워크에서 하도록 설정
  *   forward-headers-strategy: framework
  *
  * open-commons:
  *   proxy-header:
  *     real-ip: ${실제 클라이언트 IP 헤더이름}
+ *     client-port: ${실제 클라이언토 Port 헤더 이름}
  *     forwarded-for: ${프록시를 거친 모든 클라이언트 IP 목록 헤더이름}
  *     forwarded-proto: ${원 요청의 프로토콜 헤더이름}
  *     forwarded-host: ${클라이언트가 요청한 Host 헤더이름}
- *     forwarded-port: ${클라이언트가 접속한 포트 헤더이름}
+ *     forwarded-port: ${클라이언트가 요청한 Port 헤더이름}
+ *     ...
+ * </pre>
+ * 
+ * <pre>
+ * [개정이력]
+ *      날짜    	| 작성자			|	내용
+ * ------------------------------------------
+ * 2025. 7. 18.     박준홍(jhpark@ymtech.co.kr)    최초 작성
+ * 2025. 7. 30.     박준홍(jhpark@ymtech.co.kr)    추후 전달되는 헤더의 확장을 위하여 {@link HashMap}을 상속받음.
+ *                                                  * 기본 헤더 목록
+ *                                                   - real-ip
+ *                                                   - client-port
+ *                                                   - forwarded-for
+ *                                                   - forwarded-proto
+ *                                                   - forwarded-host
+ *                                                   - forwarded-port 
  * </pre>
  * 
  * @since 2025. 7. 18.
  * @version 0.8.0
  * @author parkjunhong77@gmail.com
  */
-public class HttpRequestProxyHeader {
+public class HttpRequestProxyHeader extends HashMap<String, String> {
+
+    private static final long serialVersionUID = 1L;
 
     /**
      * 실제 클라이언트 IP<br>
      * 예) Nginx의 $remote_addr : Nginx가 받은 요청의 클라이언트 IP
      */
-    private String realIp;
+    public static final String HEADER_REAL_IP = "real-ip";
     /**
      * 클라이언트가 접속한 포트<br>
      * 예) Nginx의 $remote_port : 클라이언트가 Nginx에 연결한 포트
      */
-    private String clientPort;
+    public static final String HEADER_CLIENT_PORT = "client-port";
     /**
      * 프록시를 거친 모든 클라이언트 IP 목록.<br>
      * 예) Nginx의 $proxy_add_x_forwarded_for : 기존 값 뒤에 $remote_addr를 추가
      */
-    private String forwardedFor;
+    public static final String HEADER_FORWARDED_FOR = "forwarded-for";
     /**
      * 원 요청의 프로토콜 (http 또는 https)<br>
      * 예) Nginx의 $scheme : 현재 요청 스키마 (http, https)
      */
-    private String forwardedProto;
+    public static final String HEADER_FORWARDED_PROTO = "forwarded-proto";
     /**
      * 클라이언트가 요청한 Host 헤더<br>
      * 예) Nginx의 $host : 요청한 Host
      */
-    private String forwardedHost;
+    public static final String HEADER_FORWARDED_HOST = "forwarded-host";
     /**
      * 클라이언트가 요청한 Port 헤더<br>
      * 예) Nginx의 $server_port: 요청한 port
      */
-    private String forwardedPort;
+    public static final String HEADER_FORWARDED_PORT = "forwarded-port";
 
     /**
      * <br>
@@ -122,7 +141,7 @@ public class HttpRequestProxyHeader {
      */
 
     public String getClientPort() {
-        return clientPort;
+        return get(HEADER_CLIENT_PORT);
     }
 
     /**
@@ -145,7 +164,7 @@ public class HttpRequestProxyHeader {
      */
 
     public String getForwardedFor() {
-        return forwardedFor;
+        return get(HEADER_FORWARDED_FOR);
     }
 
     /**
@@ -168,7 +187,7 @@ public class HttpRequestProxyHeader {
      */
 
     public String getForwardedHost() {
-        return forwardedHost;
+        return get(HEADER_FORWARDED_HOST);
     }
 
     /**
@@ -191,7 +210,7 @@ public class HttpRequestProxyHeader {
      */
 
     public String getForwardedPort() {
-        return forwardedPort;
+        return get(HEADER_FORWARDED_PORT);
     }
 
     /**
@@ -214,7 +233,7 @@ public class HttpRequestProxyHeader {
      */
 
     public String getForwardedProto() {
-        return forwardedProto;
+        return get(HEADER_FORWARDED_PROTO);
     }
 
     /**
@@ -237,172 +256,6 @@ public class HttpRequestProxyHeader {
      */
 
     public String getRealIp() {
-        return realIp;
+        return get(HEADER_REAL_IP);
     }
-
-    /**
-     * <br>
-     * 
-     * <pre>
-     * [개정이력]
-     *      날짜    	| 작성자	|	내용
-     * ------------------------------------------
-     * 2025. 7. 29.		박준홍			최초 작성
-     * </pre>
-     *
-     * @param clientPort
-     *            the clientPort to set
-     *
-     * @since 2025. 7. 29.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     *
-     * @see #clientPort
-     */
-    public void setClientPort(String clientRealPort) {
-        this.clientPort = clientRealPort;
-    }
-
-    /**
-     * <br>
-     * 
-     * <pre>
-     * [개정이력]
-     *      날짜    	| 작성자	|	내용
-     * ------------------------------------------
-     * 2025. 7. 18.		박준홍			최초 작성
-     * </pre>
-     *
-     * @param forwardedFor
-     *            the forwardedFor to set
-     *
-     * @since 2025. 7. 18.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     *
-     * @see #forwardedFor
-     */
-    public void setForwardedFor(@NotEmpty String forwardedFor) {
-        this.forwardedFor = forwardedFor.trim();
-    }
-
-    /**
-     * <br>
-     * 
-     * <pre>
-     * [개정이력]
-     *      날짜    	| 작성자	|	내용
-     * ------------------------------------------
-     * 2025. 7. 18.		박준홍			최초 작성
-     * </pre>
-     *
-     * @param forwardedHost
-     *            the forwardedHost to set
-     *
-     * @since 2025. 7. 18.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     *
-     * @see #forwardedHost
-     */
-    public void setForwardedHost(@NotEmpty String forwardedHost) {
-        this.forwardedHost = forwardedHost.trim();
-    }
-
-    /**
-     * <br>
-     * 
-     * <pre>
-     * [개정이력]
-     *      날짜    	| 작성자	|	내용
-     * ------------------------------------------
-     * 2025. 7. 18.		박준홍			최초 작성
-     * </pre>
-     *
-     * @param forwardedPort
-     *            the forwardedPort to set
-     *
-     * @since 2025. 7. 18.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     *
-     * @see #forwardedPort
-     */
-    public void setForwardedPort(String forwardedPort) {
-        this.forwardedPort = forwardedPort;
-    }
-
-    /**
-     * <br>
-     * 
-     * <pre>
-     * [개정이력]
-     *      날짜    	| 작성자	|	내용
-     * ------------------------------------------
-     * 2025. 7. 18.		박준홍			최초 작성
-     * </pre>
-     *
-     * @param forwardedProto
-     *            the forwardedProto to set
-     *
-     * @since 2025. 7. 18.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     *
-     * @see #forwardedProto
-     */
-    public void setForwardedProto(@NotEmpty String forwardedProto) {
-        this.forwardedProto = forwardedProto.trim();
-    }
-
-    /**
-     * <br>
-     * 
-     * <pre>
-     * [개정이력]
-     *      날짜    	| 작성자	|	내용
-     * ------------------------------------------
-     * 2025. 7. 18.		박준홍			최초 작성
-     * </pre>
-     *
-     * @param realIp
-     *            the realIp to set
-     *
-     * @since 2025. 7. 18.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     *
-     * @see #realIp
-     */
-    public void setRealIp(@NotEmpty String realIP) {
-        this.realIp = realIP.trim();
-    }
-
-    /**
-     *
-     * @since 2025. 7. 29.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     *
-     * @see java.lang.Object#toString()
-     */
-    @Override
-    public String toString() {
-        StringBuilder builder = new StringBuilder();
-        builder.append("HttpRequestProxyHeader [realIp=");
-        builder.append(realIp);
-        builder.append(", clientPort=");
-        builder.append(clientPort);
-        builder.append(", forwardedFor=");
-        builder.append(forwardedFor);
-        builder.append(", forwardedProto=");
-        builder.append(forwardedProto);
-        builder.append(", forwardedHost=");
-        builder.append(forwardedHost);
-        builder.append(", forwardedPort=");
-        builder.append(forwardedPort);
-        builder.append("]");
-        return builder.toString();
-    }
-
 }